+++ core-ng/src/main/java/core/framework/impl/web/session/SessionManager.java
@@ -2,6 +2,7 @@ package core.framework.impl.web.session;
 
 import core.framework.impl.web.request.RequestImpl;
 import core.framework.web.CookieSpec;
+import core.framework.web.Request;
 import core.framework.web.Response;
 import core.framework.web.Session;
 import org.slf4j.Logger;
@@ -9,29 +10,25 @@ import org.slf4j.LoggerFactory;
 
 import java.time.Duration;
 import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
-
-import static core.framework.util.Strings.format;
 
 /**
  * @author neo
  */
 public final class SessionManager {
     private final Logger logger = LoggerFactory.getLogger(SessionManager.class);
-    public CookieSpec sessionId;
+    private CookieSpec cookieSpec;
+    private String header;
     private Duration timeout = Duration.ofMinutes(20);
     private SessionStore store;
 
-    public SessionManager() {
-        cookie("SessionId", null);  // default domain is null, which is current host
-    }
-
-    public Session load(RequestImpl request) {
+    public Session load(Request request) {
         if (store == null) return null;  // session store is not initialized
         if (!"https".equals(request.scheme())) return null;  // only load session under https
 
         var session = new SessionImpl();
-        request.cookie(sessionId).ifPresent(sessionId -> {
+        sessionId(request).ifPresent(sessionId -> {
             logger.debug("load session");
             Map<String, String> values = store.getAndRefresh(sessionId, timeout);
             if (values != null) {
@@ -49,19 +46,32 @@ public final class SessionManager {
         if (session.invalidated && session.id != null) {
             logger.debug("invalidate session");
             store.invalidate(session.id);
-            response.cookie(sessionId, null);
+            responseSessionId(response, null);
         } else if (session.changed()) {
             logger.debug("save session");
             if (session.id == null) {
                 session.id = UUID.randomUUID().toString();
-                response.cookie(sessionId, session.id);
+                responseSessionId(response, session.id);
             }
             store.save(session.id, session.values, session.changedFields, timeout);
         }
     }
 
-    public void sessionStore(SessionStore store) {
-        if (this.store != null) throw new Error(format("session store is already configured, previous={}", this.store));
+    private Optional<String> sessionId(Request request) {
+        if (header != null) return request.header(header);
+        return request.cookie(cookieSpec);
+    }
+
+    void responseSessionId(Response response, String sessionId) {
+        if (header != null) {
+            response.header(header, sessionId == null ? "" : sessionId);
+        } else {
+            response.cookie(cookieSpec, sessionId);
+        }
+    }
+
+    public void store(SessionStore store) {
+        if (this.store != null) throw new Error("session store is already configured, previous=" + this.store);
         this.store = store;
     }
 
@@ -72,6 +82,11 @@ public final class SessionManager {
 
     public void cookie(String name, String domain) {
         if (name == null) throw new Error("name must not be null");
-        sessionId = new CookieSpec(name).domain(domain).path("/").sessionScope().httpOnly().secure();
+        cookieSpec = new CookieSpec(name).domain(domain).path("/").sessionScope().httpOnly().secure();
+    }
+
+    public void header(String name) {
+        if (name == null) throw new Error("name must not be null");
+        header = name;
     }
 }